# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QA Utils API",
    "version": "1.0.0",
    "description": "REST API for QA Utils — exposes all tool functions as HTTP endpoints. Use the Swagger UI below to explore and test every endpoint interactively, or import the OpenAPI spec into your automation framework.",
    "contact": {
      "name": "KobeT",
      "url": "https://github.com/kobenguyent/qa-utils"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development"
    }
  ],
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request body failed schema validation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Validation failed"
                },
                "details": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/agents/run": {
      "post": {
        "tags": ["Agents"],
        "summary": "Run a single autonomous agent task",
        "description": "Runs the AI agent observe-think-act loop until the task is complete or maxIterations is reached. The agent autonomously selects and calls built-in tools to complete the task.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["task", "provider"],
                "properties": {
                  "task": {
                    "type": "string",
                    "description": "Natural language task description",
                    "example": "Generate a UUID and base64-encode it",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "provider": {
                    "type": "string",
                    "enum": ["ollama", "openai", "anthropic", "google", "azure-openai"],
                    "description": "AI provider to use for the agent"
                  },
                  "apiKey": {
                    "type": "string",
                    "description": "API key (required for cloud providers)"
                  },
                  "endpoint": {
                    "type": "string",
                    "description": "Custom endpoint URL (e.g. Azure OpenAI or Ollama)"
                  },
                  "model": {
                    "type": "string",
                    "description": "Model name — uses provider default if omitted"
                  },
                  "maxIterations": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 25,
                    "default": 10,
                    "description": "Maximum observe-think-act iterations"
                  },
                  "temperature": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "default": 0.3,
                    "description": "Sampling temperature"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent run result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "answer": { "type": "string" },
                    "iterationCount": { "type": "integer" },
                    "steps": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "type": {
                            "type": "string",
                            "enum": ["thinking", "tool_call", "tool_result", "answer", "error"]
                          },
                          "content": { "type": "string" },
                          "toolName": { "type": "string" },
                          "timestamp": { "type": "integer" }
                        }
                      }
                    },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": {
            "description": "Agent execution error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/agents/orchestrate": {
      "post": {
        "tags": ["Agents"],
        "summary": "Run a multi-agent orchestration pipeline",
        "description": "Coordinates multiple specialized agents to solve a complex task.\n\n- **sequential** — each agent's output feeds into the next agent as context (chain).\n- **orchestrated** — a designated orchestrator agent breaks the task into sub-tasks, delegates them to worker agents in parallel, then synthesizes a final answer.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["task", "mode", "agents"],
                "properties": {
                  "task": {
                    "type": "string",
                    "description": "The complex task to solve",
                    "example": "Analyse this codebase for security issues and write a remediation plan",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "mode": {
                    "type": "string",
                    "enum": ["sequential", "orchestrated"],
                    "description": "Pipeline execution mode"
                  },
                  "agents": {
                    "type": "array",
                    "description": "Worker agent configurations",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": ["profileId", "name", "role", "provider"],
                      "properties": {
                        "profileId": { "type": "string" },
                        "name": { "type": "string" },
                        "role": { "type": "string" },
                        "specialty": { "type": "string" },
                        "provider": {
                          "type": "string",
                          "enum": ["ollama", "openai", "anthropic", "google", "azure-openai"]
                        },
                        "apiKey": { "type": "string" },
                        "endpoint": { "type": "string" },
                        "model": { "type": "string" },
                        "maxIterations": { "type": "integer", "minimum": 1, "maximum": 25 },
                        "temperature": { "type": "number", "minimum": 0, "maximum": 1 },
                        "order": {
                          "type": "integer",
                          "description": "Execution order for sequential mode (0-based)"
                        }
                      }
                    }
                  },
                  "orchestrator": {
                    "type": "object",
                    "description": "Orchestrator agent config — required when mode is 'orchestrated'",
                    "properties": {
                      "name": { "type": "string" },
                      "role": { "type": "string", "default": "orchestrator" },
                      "provider": {
                        "type": "string",
                        "enum": ["ollama", "openai", "anthropic", "google", "azure-openai"]
                      },
                      "apiKey": { "type": "string" },
                      "endpoint": { "type": "string" },
                      "model": { "type": "string" },
                      "maxIterations": { "type": "integer", "minimum": 1, "maximum": 25 },
                      "temperature": { "type": "number", "minimum": 0, "maximum": 1 }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pipeline execution result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "summary": { "type": "string", "description": "Final synthesized answer" },
                    "agentResults": {
                      "type": "array",
                      "description": "Per-agent outputs",
                      "items": {
                        "type": "object",
                        "properties": {
                          "agentId": { "type": "string" },
                          "agentName": { "type": "string" },
                          "success": { "type": "boolean" },
                          "answer": { "type": "string" },
                          "iterationCount": { "type": "integer" },
                          "duration": { "type": "integer", "description": "Duration in ms" }
                        }
                      }
                    },
                    "totalDuration": { "type": "integer", "description": "Total pipeline duration in ms" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": {
            "description": "Pipeline execution error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}